+++ idea/src/org/jetbrains/kotlin/idea/injection/KotlinLanguageInjector.kt
@@ -16,6 +16,7 @@
 
 package org.jetbrains.kotlin.idea.injection
 
+import com.intellij.codeInsight.AnnotationUtil
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.TextRange
@@ -27,6 +28,9 @@ import org.intellij.plugins.intelliLang.Configuration
 import org.intellij.plugins.intelliLang.inject.InjectorUtils
 import org.intellij.plugins.intelliLang.inject.config.BaseInjection
 import org.intellij.plugins.intelliLang.inject.java.JavaLanguageInjectionSupport
+import org.intellij.plugins.intelliLang.util.AnnotationUtilEx
+import org.jetbrains.kotlin.idea.util.findAnnotation
+import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.*
 import java.util.*
 
@@ -133,14 +137,35 @@ class KotlinLanguageInjector : LanguageInjector {
         val argumentIndex = (argument.parent as KtValueArgumentList).arguments.indexOf(argument)
         val psiParameter = javaMethod.parameterList.parameters.getOrNull(argumentIndex) ?: return null
 
-        return findInjection(psiParameter, Configuration.getInstance().getInjections(JavaLanguageInjectionSupport.JAVA_SUPPORT_ID))
+        val injectionInfo = findInjection(psiParameter, Configuration.getInstance().getInjections(JavaLanguageInjectionSupport.JAVA_SUPPORT_ID))
+        if (injectionInfo != null) {
+            return injectionInfo
+        }
+
+        val annotations = AnnotationUtilEx.getAnnotationFrom(
+                psiParameter,
+                Configuration.getProjectInstance(psiParameter.project).advancedConfiguration.languageAnnotationPair,
+                true)
+
+        if (annotations.size > 0) {
+            return processAnnotationInjectionInner(annotations)
+        }
+
+        return null
     }
 
     private fun injectionForKotlinCall(argument: KtValueArgument, ktFunction: KtFunction): InjectionInfo? {
         val argumentIndex = (argument.parent as KtValueArgumentList).arguments.indexOf(argument)
         val ktParameter = ktFunction.valueParameters.getOrNull(argumentIndex) ?: return null
 
-        return findInjection(ktParameter, Configuration.getInstance().getInjections(KOTLIN_SUPPORT_ID))
+        val patternInjection = findInjection(ktParameter, Configuration.getInstance().getInjections(KOTLIN_SUPPORT_ID))
+        if (patternInjection != null) {
+            return patternInjection
+        }
+
+        val injectAnnotation = ktParameter.findAnnotation(FqName(AnnotationUtil.LANGUAGE)) ?: return null
+        val languageId = extractLanguageFromInjectAnnotation(injectAnnotation) ?: return null
+        return InjectionInfo(languageId, null, null)
     }
 
     private fun findInjection(element: PsiElement?, injections: List<BaseInjection>): InjectionInfo? {
@@ -158,4 +183,12 @@ class KotlinLanguageInjector : LanguageInjector {
     }
 
     private class InjectionInfo(val languageId: String?, val prefix: String?, val suffix: String?)
+
+    private fun processAnnotationInjectionInner(annotations: Array<PsiAnnotation>): InjectionInfo? {
+        val id = AnnotationUtilEx.calcAnnotationValue(annotations, "value")
+        val prefix = AnnotationUtilEx.calcAnnotationValue(annotations, "prefix")
+        val suffix = AnnotationUtilEx.calcAnnotationValue(annotations, "suffix")
+
+        return InjectionInfo(id, prefix, suffix)
+    }
 }